requirements.txt
dlt==0.4.10
google-cloud-bigquery==3.22.0
python-dotenv==1.0.1
dbt-core==1.8.0
dbt-bigquery==1.8.0
streamlit==1.34.0
